# docs/API_SPEC_OSHC_Platform.yaml
openapi: 3.1.0
info:
  title: OSHC Comparison Platform API
  version: 0.1.0
  summary: MVP API contract for the OSHC comparison, checkout, admin, agent, and fulfilment portal.
  description: >
    Swagger/OpenAPI-compatible API specification for the OSHC comparison and purchase platform.
    This contract is intended for frontend/backend alignment, Swagger UI review, test generation,
    and implementation planning.
servers:
  - url: https://api.example.com/v1
    description: Production placeholder
  - url: https://staging-api.example.com/v1
    description: Staging placeholder
  - url: http://localhost:3000/api/v1
    description: Local development placeholder
...
security:
  - bearerAuth: []
paths:
  /quotes:
    post:
      tags: [Public Quote]
      summary: Create a quote
      description: Validates household and policy dates, calculates cover type, and creates a quote.
      security: []
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /agent/auth/login:
    post:
      tags: [Agents]
      summary: Agent login
      security: []
      operationId: agentLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /agent/quotes:
    get:
      tags: [Agents]
      summary: List agent quotes
      operationId: agentListQuotes
      responses:
        "200":
          description: Agent quotes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentQuoteListResponse"
    post:
      tags: [Agents]
      summary: Create tracked agent quote
      operationId: agentCreateQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Agent quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentQuoteResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: One or more fields are invalid.
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    example: policyEndDate
                  message:
                    type: string
                    example: Policy end date must be after policy start date.
    CoverType:
      type: string
      enum: [single, couple, single_parent_family, family]
    SalesChannel:
      type: string
      enum: [direct, agent]
    OrderStatus:
      type: string
      enum:
        - quote_created
        - application_started
        - application_completed
        - payment_pending
        - payment_failed
        - paid
        - provider_submission_pending
        - provider_submission_failed
        - certificate_pending
        - certificate_issued
        - fulfilled
        - refund_requested
        - refunded
        - cancelled
...
    CreateQuoteRequest:
      type: object
      required: [adults, children, policyStartDate, policyEndDate]
      properties:
        adults:
          type: integer
          enum: [1, 2]
        children:
          type: integer
          minimum: 0
          maximum: 10
        policyStartDate:
          type: string
          format: date
        policyEndDate:
          type: string
          format: date
        sessionId:
          type: string
          description: Optional client session ID. Server may generate one if omitted.
    Quote:
      type: object
      required: [id, adults, children, policyStartDate, policyEndDate, coverType, channel, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        adults:
          type: integer
        children:
          type: integer
        policyStartDate:
          type: string
          format: date
        policyEndDate:
          type: string
          format: date
        calculatedDurationDays:
          type: integer
        coverType:
          $ref: "#/components/schemas/CoverType"
        channel:
          $ref: "#/components/schemas/SalesChannel"
        agentId:
          type: string
          format: uuid
          nullable: true
        expiresAt:
          type: string
          format: date-time
    QuoteResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Quote"
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [accessToken, tokenType, expiresIn]
          properties:
            accessToken:
              type: string
            tokenType:
              type: string
              example: Bearer
            expiresIn:
              type: integer
              example: 3600
...
    AgentQuoteResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [quote, shareUrl]
          properties:
            quote:
              $ref: "#/components/schemas/Quote"
            shareUrl:
              type: string
              format: uri
    AgentQuoteListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              quote:
                $ref: "#/components/schemas/Quote"
              orderStatus:
                $ref: "#/components/schemas/OrderStatus"
              selectedProvider:
                type: string
                nullable: true
              certificateStatus:
                type: string
                nullable: true
...